Fixed sending DIDcomm messages with attachments (keeping attachment as links)
